pyproject.toml
[tool.poetry]
name = "teacher-backend"
version = "0.1.0"
description = ""
authors = ["avinash8419 <avihanvaisri@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
requests = "^2.32.3"
flask-sqlalchemy = "^3.1.1"
pyodbc = "^5.1.0"
pymssql = "^2.3.0"
mysql-connector-python = "^9.0.0"
python-dotenv = "^1.0.1"
pdfminer-six = "^20240706"
flask-cors = "^4.0.1"
pybase64 = "^1.3.2"
firebase-admin = "^6.5.0"
google-cloud-storage = "^2.18.0"
google-cloud = "^0.34.0"
google-api-python-client = "^2.139.0"
google = "^3.0.0"
langsmith = "^0.1.147"
langchain = "^0.3.9"
langchain-community = "^0.3.8"
langchain-google-vertexai = "^2.0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
